// MindMapper/mindmapper.client/src/app/api/model/canvasStateDto.ts
/**
 * MindMapper.Server
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CanvasStateDto { 
    id?: string | null;
    name?: string | null;
    scale?: number;
    height?: number;
    width?: number;
    seedCardId?: string | null;
}