// MindMapper/mindmapper.client/src/app/api/model/compositeCanvasStateDtoResponse.ts
/**
 * MindMapper.Server
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { CardDto } from './cardDto';
import { CanvasStateDto } from './canvasStateDto';
import { OptionDto } from './optionDto';


export interface CompositeCanvasStateDtoResponse { 
    canvas?: CanvasStateDto;
    cards?: Array<CardDto> | null;
    options?: Array<OptionDto> | null;
}