// MindMapper/mindmapper.client/src/app/api/model/optionDto.ts
/**
 * MindMapper.Server
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface OptionDto { 
    id?: string | null;
    text?: string | null;
    cardId?: string | null;
    pointToCardId?: string | null;
}